requirements.txt
onvif-zeep-async==3.12.1